adding endpoint for product

+++ models/product.models.go
@@ -6,8 +6,8 @@ type Product struct {
 	Desc       string
 	Stock      int
 	CatID      int
-	DateCreate int
-	DateUpdate int
+	DateCreate int64
+	DateUpdate int64
 }
 
 type ProductCategory struct {
